Added try catch finally so we can use it in org.webosports.app.browser as well
For debugging possible LuneOS issues like the Ajax calls not working, it's good to test it in the Browser app as well. It's beneficial if this can also be done on a desktop with Qt Creator, however that won't be able to use the PalmSystem, webos-lib and enyo-webos calls since the backend is not available. Therefore putting these in a try catch finally.

App now works properly in the LuneOS browser at least on the desktop, including the Ajax search.